Write all files into log dir

+++ cfut/__init__.py
@@ -7,10 +7,10 @@ import time
 from . import condor
 from . import slurm
 from .remote import INFILE_FMT, OUTFILE_FMT
-from .util import random_string
-from cloud import serialization
+from .util import random_string, local_filename
+import cloudpickle
 
-LOGFILE_FMT = 'cfut.log.%s.txt'
+LOGFILE_FMT = local_filename('cfut.log.%s.txt')
 
 class RemoteException(Exception):
     def __init__(self, error):
@@ -64,13 +64,15 @@ class FileWaitThread(threading.Thread):
 class ClusterExecutor(futures.Executor):
     """An abstract base class for executors that run jobs on clusters.
     """
-    def __init__(self, debug=False):
+    def __init__(self, debug=False, keep_logs=False):
+        os.makedirs(local_filename(), exist_ok=True)
         self.debug = debug
 
         self.jobs = {}
         self.job_outfiles = {}
         self.jobs_lock = threading.Lock()
         self.jobs_empty_cond = threading.Condition(self.jobs_lock)
+        self.keep_logs = keep_logs
 
         self.wait_thread = FileWaitThread(self._completion)
         self.wait_thread.start()
@@ -94,11 +96,11 @@ class ClusterExecutor(futures.Executor):
             if not self.jobs:
                 self.jobs_empty_cond.notify_all()
         if self.debug:
-            print >>sys.stderr, "job completed: %i" % jobid
+            print("job completed: %i" % jobid, file=sys.stderr)
 
-        with open(OUTFILE_FMT % workerid) as f:
+        with open(OUTFILE_FMT % workerid, 'rb') as f:
             outdata = f.read()
-        success, result = serialization.deserialize(outdata)
+        success, result = cloudpickle.loads(outdata)
 
         if success:
             fut.set_result(result)
@@ -111,19 +113,19 @@ class ClusterExecutor(futures.Executor):
 
         self._cleanup(jobid)
 
-    def submit(self, fun, *args, **kwargs):
+    def submit(self, fun, *args, additional_setup_lines=[], **kwargs):
         """Submit a job to the pool."""
         fut = futures.Future()
 
         # Start the job.
         workerid = random_string()
-        funcser = serialization.serialize((fun, args, kwargs), True)
-        with open(INFILE_FMT % workerid, 'w') as f:
+        funcser = cloudpickle.dumps((fun, args, kwargs), True)
+        with open(INFILE_FMT % workerid, 'wb') as f:
             f.write(funcser)
-        jobid = self._start(workerid)
+        jobid = self._start(workerid, additional_setup_lines)
 
         if self.debug:
-            print >>sys.stderr, "job submitted: %i" % jobid
+            print("job submitted: %i" % jobid, file=sys.stderr)
 
         # Thread will wait for it to finish.
         self.wait_thread.wait(OUTFILE_FMT % workerid, jobid)
@@ -144,12 +146,15 @@ class ClusterExecutor(futures.Executor):
 
 class SlurmExecutor(ClusterExecutor):
     """Futures executor for executing jobs on a Slurm cluster."""
-    def _start(self, workerid):
+    def _start(self, workerid, additional_setup_lines):
         return slurm.submit(
-            '{} -m cfut.remote {}'.format(sys.executable, workerid)
+            '{} -m cfut.remote {}'.format(sys.executable, workerid, additional_setup_lines=additional_setup_lines)
         )
 
     def _cleanup(self, jobid):
+        if self.keep_logs:
+            return
+
         outf = slurm.OUTFILE_FMT.format(str(jobid))
         try:
             os.unlink(outf)
@@ -162,7 +167,7 @@ class CondorExecutor(ClusterExecutor):
         super(CondorExecutor, self).__init__(debug)
         self.logfile = LOGFILE_FMT % random_string()
 
-    def _start(self, workerid):
+    def _start(self, workerid, additional_setup_lines):
         return condor.submit(sys.executable, '-m cfut.remote %s' % workerid,
                              log=self.logfile)
